Add --enable-debugging option which sets -O0 CFLAGS add compiler warnings and fix a whole bunch of them point json-c at our own copy of the repo Use the port variable in network_address_from_ip_and_port add DNS notices

+++ include/network.h
@@ -43,7 +43,7 @@ struct _NetworkAddress
     } Address;
 
     int AddressFamily;
-    int AddressLength;
+    size_t AddressLength;
 };
 
 bool network_address_from_ipstring(const char *ip, NetworkAddress *address);
